pyproject.toml
[tool.poetry]
name = "compiler_test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "<3.11,>=3.9"
pyinstaller = "4.10"
paddleocr = "2.4"
tk = "0.1.0"
paddlepaddle = "2.2.2"
setuptools = "59.8.0"
scikit-image = "0.19.1"
PyAutoGUI = "0.9.53"
keyboard = "0.13.5"
deepl = "1.4.0"
imutils = "0.5.4"
poethepoet = "0.12.2"
Pillow = "8.2.0"
opencv-python-headless = "4.5.2.54"
easyocr = "1.4.1"
googletrans = "3.1.0a0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
force-cuda11 = "pip install torch==1.10.1+cu113 torchvision==0.11.2+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html"